Refactor api into controllers
