+++ server/src/api/api.ts
@@ -1,4 +1,3 @@
-/* eslint-disable camelcase, no-param-reassign */
 import * as fs from "fs";
 import * as path from "path";
 import express from "express";
