+++ server/src/api/controllers/RequestController.ts
@@ -1,0 +1,122 @@
+/* eslint-disable camelcase */
+import { RequestStatus } from "../graphql.types";
+import { localTimestamp, onlyIfAdmin } from "../util";
+import { ItemQtyAvailable } from "./QuantityController";
+
+export interface KnexSimpleRequest {
+  request_id: number;
+  status: RequestStatus;
+  quantity: number;
+  created_at: string;
+  updated_at: string;
+}
+
+export interface KnexRequest extends KnexSimpleRequest {
+  item_id: number;
+  item_name: string;
+  description: string;
+  imageUrl: string;
+  category_id: number;
+  category_name: string;
+  totalAvailable: number;
+  maxRequestQty: number;
+  price: number;
+  hidden: boolean;
+  returnRequired: boolean;
+  approvalRequired: boolean;
+  owner: string;
+  uuid: string;
+  admin: boolean;
+  name: string;
+  email: string;
+  phone: string;
+  slackUsername: string;
+  haveID: boolean;
+  qtyInStock: number;
+  qtyUnreserved: number;
+  qtyAvailableForApproval: number;
+}
+
+export class RequestController {
+  public static redactedItem(
+    item: any,
+    isAdmin: boolean,
+    qtyInStock: ItemQtyAvailable,
+    qtyAvailableForApproval: ItemQtyAvailable,
+    qtyUnreserved: ItemQtyAvailable
+  ) {
+    return {
+      id: item.item_id,
+      item_name: item.item_name,
+      description: item.description,
+      imageUrl: item.imageUrl,
+      category: item.category_name,
+      location: {
+        location_id: item.location_id,
+        location_name: item.location_name,
+        location_hidden: item.location_hidden,
+      },
+      totalAvailable: item.totalAvailable,
+      maxRequestQty: item.maxRequestQty,
+      price: onlyIfAdmin(item.price, isAdmin),
+      hidden: item.hidden,
+      returnRequired: item.returnRequired,
+      approvalRequired: item.approvalRequired,
+      owner: onlyIfAdmin(item.owner, isAdmin),
+      qtyInStock: qtyInStock[item.item_id],
+      qtyAvailableForApproval: qtyAvailableForApproval[item.item_id],
+      qtyUnreserved: qtyUnreserved[item.item_id],
+    };
+  }
+
+  public static nestedRequest(
+    request: KnexRequest,
+    isAdmin: boolean,
+    qtyInStock: ItemQtyAvailable,
+    qtyAvailableForApproval: ItemQtyAvailable,
+    qtyUnreserved: ItemQtyAvailable
+  ) {
+    const user = {
+      uuid: request.uuid,
+      admin: request.admin,
+      name: request.name,
+      email: request.email,
+      phone: request.phone,
+      slackUsername: request.slackUsername,
+      haveID: request.haveID,
+    };
+
+    return {
+      user,
+      item: RequestController.redactedItem(
+        request,
+        isAdmin,
+        qtyInStock,
+        qtyAvailableForApproval,
+        qtyUnreserved
+      ),
+      request_id: request.request_id,
+      status: request.status,
+      location: RequestController.redactedItem(
+        request,
+        isAdmin,
+        qtyInStock,
+        qtyAvailableForApproval,
+        qtyUnreserved
+      ).location,
+      quantity: request.quantity,
+      createdAt: localTimestamp(request.created_at),
+      updatedAt: localTimestamp(request.updated_at),
+    };
+  }
+
+  public static toSimpleRequest(request: KnexSimpleRequest) {
+    return {
+      request_id: request.request_id,
+      status: request.status,
+      quantity: request.quantity,
+      createdAt: localTimestamp(request.created_at),
+      updatedAt: localTimestamp(request.updated_at),
+    };
+  }
+}
